adding pagination, need to refactor

+++ admin/includes/view_all_users.php
@@ -11,6 +11,7 @@
                             </thead>
                             <tbody>
                                <?php
+                                
                                     $query = "SELECT * FROM users ";
                                     $select_users_query = mysqli_query($connection, $query);
                                     
@@ -33,20 +34,26 @@
                                         echo "<td>{$user_lastname}</td>";
                                         echo "<td>{$user_email}</td>";
                                         echo "<td>{$user_role}</td>";
-                                        echo "<td><a href='users.php?change_to_admin={$user_id}'>Admin</a></td>";
-                                        echo "<td><a href='users.php?change_to_user={$user_id}'>User</a></td>";
-                                        echo "<td><a href='users.php?source=edit_user&user_id={$user_id}'>Edit</a></td>";
-                                        echo "<td><a href='users.php?delete={$user_id}'>Delete</a></td>";
+                                        if($_SESSION['user_role'] == 'admin'){
+                                            echo "<td><a href='users.php?change_to_admin={$user_id}'>Admin</a></td>";
+                                            echo "<td><a href='users.php?change_to_user={$user_id}'>User</a></td>";
+                                            echo "<td><a href='users.php?source=edit_user&user_id={$user_id}'>Edit</a></td>";
+                                            echo "<td><a href='users.php?delete={$user_id}'>Delete</a></td>";
+                                        }
+                                        
                                     }
                                 ?>                   
                             </tbody>
                             <?php
                             if(isset($_GET['delete']))
                             {
-                                $user_id_to_delete = $_GET['delete'];
-                                $query = "DELETE FROM users WHERE user_id = {$user_id_to_delete}" ;
-                                $delete_user_query = mysqli_query($connection, $query);
-                                header("Location: users.php");
+                                if(isset($_SESSION['user_role']) && $_SESSION['user_role'] == 'admin')
+                                {
+                                    $user_id_to_delete = mysqli_real_escape_string($_GET['delete']);
+                                    $query = "DELETE FROM users WHERE user_id = {$user_id_to_delete}" ;
+                                    $delete_user_query = mysqli_query($connection, $query);
+                                    header("Location: users.php");
+                                }
 
                             }
